# Add benchmarks and optimize html5lib a bit