--- a/benchmarks/data/wpt/random/001.html
+++ b/benchmarks/data/wpt/random/001.html
@@ -1,0 +1,3 @@
+<!doctype html>
+<title>Accessing named windows from outside the unit of related browsing contexts</title>
+<a href="001-1.html" target="test_name">Click here</a>