--- a/benchmarks/data/wpt/random/beforeunload-on-history-back-1.html
+++ b/benchmarks/data/wpt/random/beforeunload-on-history-back-1.html
@@ -1,0 +1,5 @@
+<!doctype html>
+001-1
+<script>
+addEventListener("beforeunload", function() {top.t.step(function() {top.beforeunload_fired = true})}, false);
+</script>